feat: integrate sanctuary contracts and prestige
