+++ src/game.ts
@@ -1,375 +1,142 @@
-export * from './game-live-base';
+export * from './game-sanctuary-base';
 
-import * as Base from './game-base';
-import * as Live from './game-live-base';
+import * as Base from './game-sanctuary-base';
 import {
-  journeyTierClaimKey,
-  newlyEarnedJourneyTiers,
-  seasonJourneyKey,
-} from './season-journey';
-import {
-  advanceWeeklyDirectives,
-  weeklyDirectiveKey,
-  weeklyDirectives,
-} from './weekly-directives';
-import { resolveSeasonPurchase, type SeasonShopOfferId } from './season-shop';
-import { newlyEarnedKeepsakeMilestones, seasonKeepsakeCollection } from './season-keepsakes';
-import {
-  newlyEarnedSeasonHonors,
-  type SeasonCompletionHonorId,
-} from './season-completion-honors';
-import { seasonMasteryScore } from './season-mastery-rank';
-import {
-  newlyEarnedSeasonMasteryRewards,
-  type SeasonMasteryRewardRank,
-} from './season-mastery-rewards';
-import {
-  resolveSeasonLegacyUnlock,
-  seasonLegacyNodes,
-  type SeasonLegacyNodeId,
-} from './season-legacy-board';
-import { seasonLegacyEffects } from './season-legacy-effects';
-import {
-  emptySanctuaryLevels,
-  resolveSanctuaryUpgrade,
-  sanitizeSanctuaryLevels,
-  sanctuaryEffects,
-  type SanctuaryFacilityId,
-  type SanctuaryLevels,
-} from './starlight-sanctuary';
+  advanceSanctuaryContracts,
+  sanctuaryContractSet,
+  sanctuaryPrestigeReward,
+  type SanctuaryContractId,
+  type SanctuaryContractKind,
+  type SanctuaryPrestigeRankId,
+} from './sanctuary-contracts';
 
-export type GameState = Live.GameState & {
-  claimedSeasonCompletionHonors:SeasonCompletionHonorId[];
-  claimedSeasonMasteryRanks:SeasonMasteryRewardRank[];
-  unlockedSeasonLegacyNodes:SeasonLegacyNodeId[];
-  sanctuaryLevels:SanctuaryLevels;
+export type GameState = Base.GameState & {
+  sanctuaryContractWeekKey:string|null;
+  sanctuaryContractProgress:Record<string,number>;
+  rewardedSanctuaryContracts:string[];
+  sanctuaryPrestige:number;
+  claimedSanctuaryPrestigeRanks:SanctuaryPrestigeRankId[];
 };
 
-export type Action = Live.Action
-  | { type:'PURCHASE_SEASON_OFFER'; offerId:SeasonShopOfferId }
-  | { type:'UNLOCK_SEASON_LEGACY_NODE'; nodeId:SeasonLegacyNodeId }
-  | { type:'UPGRADE_SANCTUARY'; facility:SanctuaryFacilityId };
+export type Action = Base.Action;
 
 export const initialState:GameState = {
-  ...Live.initialState,
-  claimedSeasonCompletionHonors:[],
-  claimedSeasonMasteryRanks:[],
-  unlockedSeasonLegacyNodes:[],
-  sanctuaryLevels:emptySanctuaryLevels(),
+  ...Base.initialState,
+  sanctuaryContractWeekKey:null,
+  sanctuaryContractProgress:{},
+  rewardedSanctuaryContracts:[],
+  sanctuaryPrestige:0,
+  claimedSanctuaryPrestigeRanks:[],
 };
 
-const honorIds:SeasonCompletionHonorId[] = ['first_complete','four_seasons','perfect_year','eight_complete'];
-const masteryRewardRanks:SeasonMasteryRewardRank[] = ['traveler','chronicler','guardian','eternal'];
-const legacyNodeIds = seasonLegacyNodes.map(node => node.id);
-const isRecord = (value:unknown): value is Record<string,unknown> => typeof value === 'object' && value !== null && !Array.isArray(value);
+const contractIds:SanctuaryContractId[] = ['training_focus','field_patrol','warm_bond','guardian_sortie'];
+const prestigeRanks:{ id:Exclude<SanctuaryPrestigeRankId,'outpost'>; threshold:number }[] = [
+  { id:'haven', threshold:20 },
+  { id:'sanctum', threshold:50 },
+  { id:'citadel', threshold:100 },
+  { id:'celestial', threshold:180 },
+];
+const weekKeyPattern = /^\d+-(?:[1-9]|1[0-2])-[1-4]$/;
+const rewardKeyPattern = /^\d+-(?:[1-9]|1[0-2])-[1-4]:(training_focus|field_patrol|warm_bond|guardian_sortie)$/;
+const isRecord = (value:unknown):value is Record<string,unknown> => typeof value === 'object' && value !== null && !Array.isArray(value);
+const safeInt = (value:unknown) => typeof value === 'number' && Number.isFinite(value) ? Math.max(0,Math.floor(value)) : 0;
 
-function hydrateSeasonCompletionHonors(raw:unknown):SeasonCompletionHonorId[] {
-  if (!Array.isArray(raw)) return [];
-  return [...new Set(raw.filter((value):value is SeasonCompletionHonorId => typeof value === 'string' && honorIds.includes(value as SeasonCompletionHonorId)))];
+function sanitizeContractProgress(raw:unknown):Record<string,number> {
+  if (!isRecord(raw)) return {};
+  const result:Record<string,number> = {};
+  for (const id of contractIds) if (Object.prototype.hasOwnProperty.call(raw,id)) result[id] = safeInt(raw[id]);
+  return result;
 }
 
-function hydrateSeasonMasteryRanks(raw:unknown):SeasonMasteryRewardRank[] {
+function sanitizeRewardedContracts(raw:unknown):string[] {
   if (!Array.isArray(raw)) return [];
-  return [...new Set(raw.filter((value):value is SeasonMasteryRewardRank => typeof value === 'string' && masteryRewardRanks.includes(value as SeasonMasteryRewardRank)))];
+  return [...new Set(raw.filter((value):value is string => typeof value === 'string' && rewardKeyPattern.test(value)))];
 }
 
-function hydrateSeasonLegacyNodes(raw:unknown):SeasonLegacyNodeId[] {
+function sanitizePrestigeRanks(raw:unknown):SanctuaryPrestigeRankId[] {
+  const ids:SanctuaryPrestigeRankId[] = ['haven','sanctum','citadel','celestial'];
   if (!Array.isArray(raw)) return [];
-  return [...new Set(raw.filter((value):value is SeasonLegacyNodeId => typeof value === 'string' && legacyNodeIds.includes(value as SeasonLegacyNodeId)))];
+  return [...new Set(raw.filter((value):value is SanctuaryPrestigeRankId => typeof value === 'string' && ids.includes(value as SanctuaryPrestigeRankId)))];
 }
 
 export function hydrateGameState(raw:unknown):GameState {
   const source = isRecord(raw) ? raw : {};
   return {
-    ...Live.hydrateGameState(raw),
-    claimedSeasonCompletionHonors:hydrateSeasonCompletionHonors(source.claimedSeasonCompletionHonors),
-    claimedSeasonMasteryRanks:hydrateSeasonMasteryRanks(source.claimedSeasonMasteryRanks),
-    unlockedSeasonLegacyNodes:hydrateSeasonLegacyNodes(source.unlockedSeasonLegacyNodes),
-    sanctuaryLevels:sanitizeSanctuaryLevels(source.sanctuaryLevels),
+    ...Base.hydrateGameState(raw),
+    sanctuaryContractWeekKey:typeof source.sanctuaryContractWeekKey === 'string' && weekKeyPattern.test(source.sanctuaryContractWeekKey) ? source.sanctuaryContractWeekKey : null,
+    sanctuaryContractProgress:sanitizeContractProgress(source.sanctuaryContractProgress),
+    rewardedSanctuaryContracts:sanitizeRewardedContracts(source.rewardedSanctuaryContracts),
+    sanctuaryPrestige:safeInt(source.sanctuaryPrestige),
+    claimedSanctuaryPrestigeRanks:sanitizePrestigeRanks(source.claimedSanctuaryPrestigeRanks),
   };
 }
 
-function preserveSeasonMeta(state:GameState, next:Live.GameState):GameState {
-  if (next === state) return state;
+function persistentMeta(state:GameState) {
   return {
-    ...next,
-    claimedSeasonCompletionHonors:state.claimedSeasonCompletionHonors ?? [],
-    claimedSeasonMasteryRanks:state.claimedSeasonMasteryRanks ?? [],
-    unlockedSeasonLegacyNodes:state.unlockedSeasonLegacyNodes ?? [],
-    sanctuaryLevels:state.sanctuaryLevels ?? emptySanctuaryLevels(),
+    sanctuaryContractWeekKey:state.sanctuaryContractWeekKey ?? null,
+    sanctuaryContractProgress:state.sanctuaryContractProgress ?? {},
+    rewardedSanctuaryContracts:state.rewardedSanctuaryContracts ?? [],
+    sanctuaryPrestige:state.sanctuaryPrestige ?? 0,
+    claimedSanctuaryPrestigeRanks:state.claimedSanctuaryPrestigeRanks ?? [],
   };
 }
 
-function applySeasonCompletionHonors(previous:GameState, next:GameState):GameState {
-  const claimed = previous.claimedSeasonCompletionHonors ?? [];
-  const earned = newlyEarnedSeasonHonors(next.seasonJourneyHistory,claimed);
-  if (!earned.length) return next;
-  return {
-    ...next,
-    gold:next.gold + earned.reduce((sum,item) => sum + item.reward.gold,0),
-    gems:next.gems + earned.reduce((sum,item) => sum + item.reward.gems,0),
-    claimedSeasonCompletionHonors:[...claimed,...earned.map(item => item.id)],
-  };
+function actionKind(action:Action):SanctuaryContractKind|null {
+  if (action.type === 'FINISH_TRAINING') return 'training';
+  if (action.type === 'GO_OUTING') return 'outing';
+  if (action.type === 'GIVE_GIFT') return 'gift';
+  if (action.type === 'FINISH_EXPEDITION_STAGE') return 'expedition';
+  return null;
 }
 
-function applySeasonMasteryRewards(previous:GameState, next:GameState):GameState {
-  const claimed = previous.claimedSeasonMasteryRanks ?? [];
-  const keepsakes = seasonKeepsakeCollection(next.seasonShopPurchases).total;
-  const completedSeasons = next.seasonJourneyHistory.filter(entry => entry.tiersCompleted >= 10).length;
-  const honors = next.claimedSeasonCompletionHonors.length;
-  const score = seasonMasteryScore({ completedSeasons, keepsakes, honors });
-  const earned = newlyEarnedSeasonMasteryRewards(score,claimed);
-  if (!earned.length) return { ...next, claimedSeasonMasteryRanks:claimed };
-  return {
-    ...next,
-    gold:next.gold + earned.reduce((sum,item) => sum + item.reward.gold,0),
-    gems:next.gems + earned.reduce((sum,item) => sum + item.reward.gems,0),
-    claimedSeasonMasteryRanks:[...claimed,...earned.map(item => item.rank)],
-  };
-}
-
-function applySanctuaryTraining(previous:GameState,next:GameState):GameState {
-  const effects = sanctuaryEffects(previous.sanctuaryLevels ?? emptySanctuaryLevels());
-  let result = next;
-  if (effects.trainingPercent > 0) {
-    const stats = { ...next.stats };
-    const growthStats = ['strength','intelligence','magic','morality'] as const;
-    for (const key of growthStats) {
-      const delta = Math.max(0,next.stats[key] - previous.stats[key]);
-      if (delta > 0) stats[key] = Math.min(100,next.stats[key] + delta * effects.trainingPercent / 100);
-    }
-    result = { ...result, stats };
-  }
-  const grade = Base.trainingGrade(previous.trainingScore);
-  const masteryBonus = effects.masteryAllMonth || ((grade === 'A' || grade === 'S') ? effects.masteryStrongMonth : 0);
-  if (masteryBonus > 0) {
-    const ids = Object.keys(next.mastery) as Array<keyof typeof next.mastery>;
-    const target = ids
-      .filter(id => next.mastery[id].xp > previous.mastery[id].xp)
-      .sort((a,b) => (next.mastery[b].xp - previous.mastery[b].xp) - (next.mastery[a].xp - previous.mastery[a].xp))[0];
-    if (target) result = { ...result, mastery:{ ...result.mastery, [target]:{ ...result.mastery[target], xp:result.mastery[target].xp + 1 } } };
-  }
-  return result;
-}
-
-function applySanctuaryRecovery(state:GameState):GameState {
-  const effects = sanctuaryEffects(state.sanctuaryLevels ?? emptySanctuaryLevels());
-  if (!effects.fatigueRecovery && !effects.stressRecovery) return state;
-  const stats = {
-    ...state.stats,
-    fatigue:Math.max(0,state.stats.fatigue - effects.fatigueRecovery),
-    stress:Math.max(0,state.stats.stress - effects.stressRecovery),
-  };
-  return { ...state, stats, condition:Base.deriveCondition(stats) };
-}
-
-function applyJourneyBonus(next:GameState,key:ReturnType<typeof seasonJourneyKey>,bonus:number,updateArchive:boolean):GameState {
-  if (bonus <= 0) return next;
-  const before = next.seasonJourneyScores[key] ?? 0;
-  const score = before + bonus;
-  const earned = newlyEarnedJourneyTiers(before,score,next.claimedSeasonJourneyTiers,key);
-  const claims = [...next.claimedSeasonJourneyTiers];
+function applyPrestigeRewards(state:GameState,prestige:number,claimed:SanctuaryPrestigeRankId[]) {
+  const nextClaimed = [...claimed];
   let gold = 0;
   let gems = 0;
-  let tokens = 0;
-  for (const tier of earned) {
-    claims.push(journeyTierClaimKey(key,tier.tier));
-    gold += tier.reward.gold;
-    gems += tier.reward.gems;
-    tokens += tier.reward.tokens;
+  for (const rank of prestigeRanks) {
+    if (prestige < rank.threshold || nextClaimed.includes(rank.id)) continue;
+    const reward = sanctuaryPrestigeReward(rank.id);
+    nextClaimed.push(rank.id);
+    gold += reward.gold;
+    gems += reward.gems;
   }
-  const tiersCompleted = claims.filter(claim => claim.startsWith(`${key}:`)).length;
-  const history = updateArchive ? next.seasonJourneyHistory.map(entry => entry.key === key ? {
-    ...entry,
-    score,
-    tiersCompleted:Math.max(entry.tiersCompleted,tiersCompleted),
-    tokensEarned:entry.tokensEarned + tokens,
-  } : entry) : next.seasonJourneyHistory;
+  return { claimed:nextClaimed, gold, gems };
+}
+
+function applyContractAction(previous:GameState,next:GameState,kind:SanctuaryContractKind):GameState {
+  const key = `${previous.year}-${previous.month}-${previous.week}`;
+  const progress = previous.sanctuaryContractWeekKey === key ? (previous.sanctuaryContractProgress ?? {}) : {};
+  const contracts = sanctuaryContractSet(previous.year,previous.month,previous.week,previous.sanctuaryLevels);
+  const rewarded = previous.rewardedSanctuaryContracts ?? [];
+  const completedThisWeek = rewarded
+    .filter(item => item.startsWith(`${key}:`))
+    .map(item => item.slice(key.length + 1));
+  const advanced = advanceSanctuaryContracts(contracts,progress,{ kind },completedThisWeek);
+  const newlyCompleted = advanced.completed.filter(item => !rewarded.includes(`${key}:${item.id}`));
+  const prestigeGain = newlyCompleted.reduce((sum,item) => sum + item.prestige,0);
+  const prestige = (previous.sanctuaryPrestige ?? 0) + prestigeGain;
+  const claimed = previous.claimedSanctuaryPrestigeRanks ?? [];
+  const rankRewards = applyPrestigeRewards(previous,prestige,claimed);
+  if (!newlyCompleted.length && previous.sanctuaryContractWeekKey === key && advanced.progress === progress) return next;
   return {
     ...next,
-    seasonJourneyScores:{ ...next.seasonJourneyScores, [key]:score },
-    claimedSeasonJourneyTiers:claims,
-    seasonTokenBalances:{ ...next.seasonTokenBalances, [key]:(next.seasonTokenBalances[key] ?? 0) + tokens },
-    seasonJourneyHistory:history,
-    gold:next.gold + gold,
-    gems:next.gems + gems,
-    lastLiveOpsProgress:next.lastLiveOpsProgress ? {
-      ...next.lastLiveOpsProgress,
-      journeyPoints:next.lastLiveOpsProgress.journeyPoints + bonus,
-      seasonTiersClaimed:[...next.lastLiveOpsProgress.seasonTiersClaimed,...earned.map(item => item.tier)],
-      tokensEarned:next.lastLiveOpsProgress.tokensEarned + tokens,
-    } : next.lastLiveOpsProgress,
+    sanctuaryContractWeekKey:key,
+    sanctuaryContractProgress:advanced.progress,
+    rewardedSanctuaryContracts:[...rewarded,...newlyCompleted.map(item => `${key}:${item.id}`)],
+    sanctuaryPrestige:prestige,
+    claimedSanctuaryPrestigeRanks:rankRewards.claimed,
+    gold:next.gold + newlyCompleted.reduce((sum,item) => sum + item.reward.gold,0) + rankRewards.gold,
+    gems:next.gems + newlyCompleted.reduce((sum,item) => sum + item.reward.gems,0) + rankRewards.gems,
   };
 }
 
-function applyExpeditionMetaBonuses(previous:GameState,next:GameState):GameState {
-  if (!next.lastExpeditionResult?.accepted) return next;
-  const sanctuaryBonus = sanctuaryEffects(previous.sanctuaryLevels ?? emptySanctuaryLevels()).expeditionJourneyBonus;
-  const legacyBonus = seasonLegacyEffects(previous.unlockedSeasonLegacyNodes ?? []).expeditionJourneyBonus;
-  return applyJourneyBonus(next,seasonJourneyKey(previous.year,previous.month),sanctuaryBonus + legacyBonus,false);
-}
-
-function applyMonthlyLegacyJourney(previous:GameState,next:GameState):GameState {
-  const bonus = seasonLegacyEffects(previous.unlockedSeasonLegacyNodes ?? []).monthlyJourneyBonus;
-  return applyJourneyBonus(next,seasonJourneyKey(previous.year,previous.month),bonus,true);
-}
-
-export function reducer(state:GameState, action:Action):GameState {
+export function reducer(state:GameState,action:Action):GameState {
   if (action.type === 'RESET') return initialState;
-
-  if (action.type === 'UPGRADE_SANCTUARY') {
-    const result = resolveSanctuaryUpgrade({
-      facility:action.facility,
-      levels:state.sanctuaryLevels ?? emptySanctuaryLevels(),
-      gold:state.gold,
-      materials:state.expeditionMaterials,
-      renown:state.regionalRenown,
-    });
-    if (!result.accepted) return state;
-    return {
-      ...state,
-      gold:state.gold - result.cost.gold,
-      expeditionMaterials:{
-        star_bark:state.expeditionMaterials.star_bark - result.cost.materials.star_bark,
-        arcane_shard:state.expeditionMaterials.arcane_shard - result.cost.materials.arcane_shard,
-        wind_pearl:state.expeditionMaterials.wind_pearl - result.cost.materials.wind_pearl,
-      },
-      sanctuaryLevels:{ ...state.sanctuaryLevels, [action.facility]:result.nextLevel },
-    };
-  }
-
-  if (action.type === 'UNLOCK_SEASON_LEGACY_NODE') {
-    const result = resolveSeasonLegacyUnlock({
-      nodeId:action.nodeId,
-      history:state.seasonJourneyHistory,
-      honors:state.claimedSeasonCompletionHonors,
-      unlocked:state.unlockedSeasonLegacyNodes ?? [],
-    });
-    if (!result.accepted) return state;
-    return {
-      ...state,
-      gold:state.gold + result.reward.gold,
-      gems:state.gems + result.reward.gems,
-      unlockedSeasonLegacyNodes:result.unlocked,
-    };
-  }
-
-  if (action.type === 'PURCHASE_SEASON_OFFER') {
-    const journeyKey = seasonJourneyKey(state.year,state.month);
-    const result = resolveSeasonPurchase({
-      seasonKey:journeyKey,
-      offerId:action.offerId,
-      tokens:state.seasonTokenBalances[journeyKey] ?? 0,
-      purchaseKeys:state.seasonShopPurchases,
-    });
-    if (!result.accepted) return state;
-    const inventory = { ...state.inventory };
-    for (const [id,amount] of Object.entries(result.reward.inventory)) {
-      const key = id as keyof typeof inventory;
-      inventory[key] += amount ?? 0;
-    }
-    const expeditionMaterials = { ...state.expeditionMaterials };
-    for (const [id,amount] of Object.entries(result.reward.materials)) {
-      const key = id as keyof typeof expeditionMaterials;
-      expeditionMaterials[key] += amount ?? 0;
-    }
-    const seasonShopPurchases = [...state.seasonShopPurchases,result.purchaseKey];
-    const claimedMilestones = state.claimedSeasonKeepsakeMilestones ?? [];
-    const keepsakeMilestones = newlyEarnedKeepsakeMilestones(seasonShopPurchases,claimedMilestones);
-    const keepsakeGold = keepsakeMilestones.reduce((sum,item) => sum + item.reward.gold,0);
-    const keepsakeGems = keepsakeMilestones.reduce((sum,item) => sum + item.reward.gems,0);
-    const purchased:GameState = {
-      ...state,
-      gold:state.gold + result.reward.gold + keepsakeGold,
-      gems:state.gems + keepsakeGems,
-      inventory,
-      expeditionMaterials,
-      seasonTokenBalances:{ ...state.seasonTokenBalances, [journeyKey]:result.tokens },
-      seasonShopPurchases,
-      claimedSeasonKeepsakeMilestones:[...claimedMilestones,...keepsakeMilestones.map(item => item.id)],
-      claimedSeasonMasteryRanks:state.claimedSeasonMasteryRanks ?? [],
-      unlockedSeasonLegacyNodes:state.unlockedSeasonLegacyNodes ?? [],
-      sanctuaryLevels:state.sanctuaryLevels ?? emptySanctuaryLevels(),
-    };
-    return applySeasonMasteryRewards(state,purchased);
-  }
-
-  if (action.type !== 'FINISH_TRAINING') {
-    let liveNext = Live.reducer(state,action);
-    if (liveNext === state) return state;
-    if (action.type === 'NEXT_MONTH') liveNext = applyMonthlyLegacyJourney(state,preserveSeasonMeta(state,liveNext));
-    let preserved = preserveSeasonMeta(state,liveNext);
-    if (action.type === 'FINISH_EXPEDITION_STAGE') preserved = applyExpeditionMetaBonuses(state,preserved);
-    if (action.type !== 'NEXT_MONTH') return preserved;
-    preserved = applySanctuaryRecovery(preserved);
-    const honored = applySeasonCompletionHonors(state,preserved);
-    return applySeasonMasteryRewards(state,honored);
-  }
-
-  const baseNext = Base.reducer(state,action as Base.Action);
+  const baseNext = Base.reducer(state,action);
   if (baseNext === state) return state;
-  const rawNext:GameState = {
-    ...state,
-    ...baseNext,
-    claimedSeasonCompletionHonors:state.claimedSeasonCompletionHonors ?? [],
-    claimedSeasonMasteryRanks:state.claimedSeasonMasteryRanks ?? [],
-    unlockedSeasonLegacyNodes:state.unlockedSeasonLegacyNodes ?? [],
-    sanctuaryLevels:state.sanctuaryLevels ?? emptySanctuaryLevels(),
-  };
-  const next = applySanctuaryTraining(state,rawNext);
-  const weekKey = weeklyDirectiveKey(state.year,state.month,state.week);
-  const directives = weeklyDirectives(state.year,state.month,state.week);
-  const progress = state.weeklyDirectiveKey === weekKey ? state.weeklyDirectiveProgress : {};
-  const weekly = advanceWeeklyDirectives(
-    directives,
-    progress,
-    { kind:'training' },
-    state.rewardedWeeklyDirectives,
-    weekKey,
-  );
-  const journeyKey = seasonJourneyKey(state.year,state.month);
-  const previousScore = state.seasonJourneyScores[journeyKey] ?? 0;
-  const nextScore = previousScore + weekly.reward.journeyPoints;
-  const earnedTiers = newlyEarnedJourneyTiers(previousScore,nextScore,state.claimedSeasonJourneyTiers,journeyKey);
-  const claims = [...state.claimedSeasonJourneyTiers];
-  const rewardedWeekly = [...state.rewardedWeeklyDirectives];
-  let gold = 0;
-  let gems = 0;
-  let tierTokens = 0;
-  for (const tier of earnedTiers) {
-    claims.push(journeyTierClaimKey(journeyKey,tier.tier));
-    gold += tier.reward.gold;
-    gems += tier.reward.gems;
-    tierTokens += tier.reward.tokens;
-  }
-  for (const directive of weekly.completed) {
-    const rewardKey = `${weekKey}:${directive.id}`;
-    if (!rewardedWeekly.includes(rewardKey)) rewardedWeekly.push(rewardKey);
-  }
-  const legacyWeeklyTokens = weekly.completed.length * seasonLegacyEffects(state.unlockedSeasonLegacyNodes ?? []).weeklyTokenBonus;
-  const tokensEarned = tierTokens + weekly.reward.tokens + legacyWeeklyTokens;
-  return {
-    ...next,
-    seasonJourneyScores:{ ...state.seasonJourneyScores, [journeyKey]:nextScore },
-    claimedSeasonJourneyTiers:claims,
-    seasonTokenBalances:{ ...state.seasonTokenBalances, [journeyKey]:(state.seasonTokenBalances[journeyKey] ?? 0) + tokensEarned },
-    weeklyDirectiveKey:weekKey,
-    weeklyDirectiveProgress:weekly.progress,
-    rewardedWeeklyDirectives:rewardedWeekly,
-    seasonJourneyHistory:state.seasonJourneyHistory,
-    seasonShopPurchases:state.seasonShopPurchases,
-    claimedSeasonKeepsakeMilestones:state.claimedSeasonKeepsakeMilestones ?? [],
-    claimedSeasonCompletionHonors:state.claimedSeasonCompletionHonors ?? [],
-    claimedSeasonMasteryRanks:state.claimedSeasonMasteryRanks ?? [],
-    unlockedSeasonLegacyNodes:state.unlockedSeasonLegacyNodes ?? [],
-    sanctuaryLevels:state.sanctuaryLevels ?? emptySanctuaryLevels(),
-    gold:next.gold + gold,
-    gems:next.gems + gems,
-    lastLiveOpsProgress:{
-      journeyPoints:weekly.reward.journeyPoints,
-      seasonTiersClaimed:earnedTiers.map(tier => tier.tier),
-      weeklyCompleted:weekly.completed.map(item => item.id),
-      tokensEarned,
-    },
-  };
+  let next:GameState = { ...baseNext, ...persistentMeta(state) };
+  const kind = actionKind(action);
+  if (!kind) return next;
+  if (kind === 'expedition' && !next.lastExpeditionResult?.accepted) return next;
+  return applyContractAction(state,next,kind);
 }
